backend/django_app/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.6
django-cors-headers>=4.4,<5.0
django-environ>=0.11,<0.12
django-storages[s3]>=1.14,<2.0
psycopg>=3.2,<3.3
psycopg-binary>=3.2,<3.3
redis>=5.0,<6.0
celery[redis]>=5.4,<5.5
gunicorn>=22.0,<24.0
Pillow>=10.0,<12.0
drf-spectacular>=0.27,<0.28
django-unfold>=0.86,<0.87
django-unfold>=0.86,<0.87